// src/types/openai-types.ts
export namespace openai {
	export interface CreateChatCompletionDeltaResponse {
		id: string
		object: 'chat.completion.chunk'
		created: number
		model: string
		choices: [
			{
				delta: {
					role: Role
					content?: string
				}
				index: number
				finish_reason: string | null
			}
		]
	}
// ...
}